+++ src/App.js
@@ -4,7 +4,7 @@ import Hidden from '@material-ui/core/Hidden';
 import "./index.css";
 // importing pages 
 import Index from './pages/home/index'
-import Team from './pages/team'
+import Team from './pages/team/index'
 import Hacktoberfest from './pages/hacktoberfest'
 import Events from './pages/events'
 import Projects from './pages/projects'
