+++ src/components/navbar/navbar.jsx
@@ -1,16 +1,26 @@
 import React from 'react';
 import Toolbar from '@material-ui/core/Toolbar';
 import Button from '@material-ui/core/Button';
-
+import { useHistory } from "react-router-dom";
 
 const Navbar = () => {
-    return(
+    const history = useHistory();
+    const router = (path) => {
+        history.push(path);
+    };
+    const paths = [
+        { name: "About Us", path: "/" },
+        { name: "Our Story", path: "/team" },
+        { name: "Events", path: "/events" },
+        { name: "Projects", path: "/projects" }
+    ];
+    return (
         <Toolbar>
-            <Button color="inherit">About Us</Button>
-            <Button color="inherit">Projects</Button>
-            <Button color="inherit">Our Story</Button>
-            <Button color="inherit">Events</Button>
+            {
+            paths.map(el => <Button onClick={()=>{ router(el.path)}}>{el.name}</Button>)
+            }          
             <Button variant="outlined" border="true" color="secondary">Join Us</Button>
         </Toolbar>
-   ) }
+    )
+}
 export default Navbar
