+++ src/pages/team/index.jsx
@@ -1,13 +1,23 @@
 import React from 'react';
 
 
-const Index = props => {
+import DisplayTeams from './coremembers/index';
+import { team } from '../../data/airtable'
+import OurStory from './ourstory/index';
 
+
+
+
+const Team = () => {
+
+    const data = team.useTeam()
+    console.log(data)
     return (
-        <div>
-
-        </div>
+        <>
+            <OurStory/>
+       
+            <DisplayTeams members={data} />
+        </>
     )
 }
-
-export default Index
+export default Team;
